// src/app/types/enums.ts
export enum AccountType {
    Checking = 'Checking',
    Savings = 'Savings',
}

export enum FormResponseNames {
    AccountNumber = 'accountnumber',
    ConfirmAccountNumber = 'confirmaccountnumber',
    RoutingNumber = 'routingnumber',
    AccountType = 'accounttype',
    PaymentAmount = 'paymentamount',
    AccountPaymentPrefix = 'account_',
}